CORE-10593 Renamed 'configuration-rpc-ops-impl' to 'configuration-rest-resource-impl' (#3202)
* Renamed 'configuration-rpc-ops-impl' to 'configuration-rest-resource-impl'

* Renaming 'configuration-rpcops-service-impl' to 'configuration-rest-resource-service-impl' in test.bndrun
