+++ components/configuration/configuration-rest-resource-service-impl/build.gradle
