+++ components/configuration/configuration-rest-resource-service-impl/detekt-baseline.xml
