+++ components/configuration/configuration-rest-resource-service-impl/src/main/kotlin/net/corda/configuration/restresource/impl/Constants.kt
@@ -1,4 +1,4 @@
-package net.corda.configuration.rpcops.impl
+package net.corda.configuration.rest.impl
 
 internal const val GROUP_NAME = "config.management"
 internal const val CLIENT_NAME_HTTP = "config.manager.http"
