+++ components/configuration/configuration-rest-resource-service-impl/src/main/kotlin/net/corda/configuration/restresource/impl/exception/ConfigException.kt
@@ -1,4 +1,4 @@
-package net.corda.configuration.rpcops.impl.exception
+package net.corda.configuration.rest.impl.exception
 
 import net.corda.data.config.ConfigurationSchemaVersion
 import net.corda.httprpc.ResponseCode
