+++ components/configuration/configuration-rest-resource-service-impl/src/main/kotlin/net/corda/configuration/restresource/impl/exception/ConfigRestResourceException.kt
@@ -1,0 +1,6 @@
+package net.corda.configuration.rest.impl.exception
+
+import net.corda.v5.base.exceptions.CordaRuntimeException
+
+/** Exceptions related to the [net.corda.configuration.rpcops.impl.v1.ConfigRestResourceImpl]. */
+class ConfigRestResourceException(message: String, e: Exception? = null) : CordaRuntimeException(message, e)
