+++ components/configuration/configuration-rest-resource-service-impl/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
