+++ settings.gradle
@@ -73,7 +73,7 @@ include 'applications:workers:release:p2p-link-manager-worker'
 include 'applications:workers:release:rest-worker'
 include 'applications:workers:worker-common'
 include 'applications:workers:workers-smoketest'
-include 'components:configuration:configuration-rpcops-service-impl'
+include 'components:configuration:configuration-rest-resource-service-impl'
 include 'components:configuration:configuration-read-service'
 include 'components:configuration:configuration-read-service-impl'
 include 'components:configuration:configuration-write-service'
